+++ CHANGES.md
@@ -2,6 +2,7 @@ Master
 ------------------
 
 * atdgen: use Yojson 2.0 API (#299)
+* atdpy: Support recursive definitions (#315)
 * atdts: fix nullable object field writer (#312)
 
 2.9.1 (2022-06-10)